# backend/src/main/resources/api.yml
openapi: 3.1.0
info:
  title: 'Nerdy Blog API'
  description: 'API for a blog website with mainly computer science and economics topics.'
  version: 0.0.1
servers:
  - url: https://{userName}.{domainName}:{port}/blog/{version}
    variables:
      userName:
        default: nerdy
      port:
        enum:
          - "443"
          - "8443"
        default: "443"
      version:
        default: v1
      domainName:
        default: dev
...
paths:
  /articles:
    get:
      tags: [articles]
      summary: Get all articles for reading
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  recentArticles:
                    $ref: '#/components/schemas/articleList'
                  mostViewedArticles:
                    $ref: '#/components/schemas/articleList'
                  featuredArticles:
                    $ref: '#/components/schemas/articleList'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /articles/{id}:
    get:
      tags: [articles]
      summary: Read an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/articleDetail'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /tags/{id}:
    delete:
      tags: [tags]
      summary: Delete a tag
      security:
        - bearerHttpAuthentication:
            - write: admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "OK"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    patch:
      tags: [tags]
      summary: Change information of a tag
      security:
        - bearerHttpAuthentication:
            - write: authorizedUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tagPatchRequest'
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tagAdminDetail'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  schemas:
    category:
      description: Category of a blog
      type: string
      enum: ['cs', 'econs']
    tag:
      description: Tag of a blog
      type: object
      required: [slugName, category]
      properties:
        slugName:
          type: string
          maxLength: 255
        category:
          $ref: '#/components/schemas/category'
...
    tagPatchRequest:
      type: object
      properties:
        slugName:
          type: string
          maxLength: 255
        categoryId:
          type: integer
          format: int64
        isActive:
          type: boolean
    tagAdminDetail:
      type: object
      required: [tagId, slugName, isActive, category]
      properties:
        tagId:
          type: integer
          format: int64
        slugName:
          type: string
          maxLength: 255
        isActive:
          type: boolean
        category:
          $ref: '#/components/schemas/category'
    articleShow:
      description: Article showcase
      type: object
      properties:
        name:
          type: string
        author:
          type: string
        uploadAt:
          type: string
          format: date
        upvotes:
          type: integer
        downvote:
          type: integer
        category:
          $ref: '#/components/schemas/category'
        tags:
          type: array
          minItems: 1
          maxItems: 5
          items:
            $ref: '#/components/schemas/tag'
    articleDetail:
      description: Content of article
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        author:
          type: string
        uploadAt:
          type: string
          format: date
        upvotes:
          type: integer
        downvote:
          type: integer
        category:
          $ref: '#/components/schemas/category'
        tags:
          type: array
          minItems: 1
          maxItems: 5
          items:
            $ref: '#/components/schemas/tag'
...
    articleList:
      type: array
      items:
        $ref: '#/components/schemas/articleShow'
...
    error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  securitySchemes:
    bearerHttpAuthentication:
      description: Bearer token using JWT
      type: http
      scheme: Bearer
      bearerFormat: JWT